Split function to math utils
