+++ src/math_utils.cpp
@@ -5,3 +5,15 @@
  * @author Amro Al-Baali (albaalia@live.com)
  * @date 2022-Mar-16
  */
+
+#include "turtle_nav_cpp/math_utils.hpp"
+
+namespace turtle_nav_cpp
+{
+bool IsPerfectSquare(double x)
+{
+  double x_sqrt = sqrt(x);
+  return (x_sqrt - floor(x_sqrt)) == 0;
+}
+
+}  // namespace turtle_nav_cpp
